:sparkles: split goml loader
